--- src/backend/.env.example
@@ -1,2 +1,0 @@
-DATA_DIR="../../data"
-PLUGIN_DIR="../../notebooks/"
