--- src/backend/app/__init__.py
@@ -1,3 +1,0 @@
-from app.main import app
-
-__all__ = ["app"]
