+++ src/backend/ocelescope-backend/pyproject.toml
@@ -1,0 +1,31 @@
+[project]
+name = "ocelescope-backend"
+version = "0.1.0"
+description = "Ocelescope backend host application"
+requires-python = ">=3.13"
+dependencies = [
+  "pm4py~=2.7.22",
+  "pint>=0.23,<0.24",
+  "pandas>=2.3.1",
+  "numpy>=2.3.2",
+  "cachetools>=6.1.0",
+  "pydantic>=2.11.7",
+  "pydantic-settings>=2.2.1,<3",
+  "fastapi[standard]>=0.136.0",
+  "uvicorn>=0.28.0,<0.29",
+  "ocelescope",
+  "orjson>=3.11.3",
+]
+
+[project.scripts]
+ocelescope-backend = "ocelescope_backend.cli:main"
+
+[build-system]
+requires = ["uv_build>=0.8.4,<0.9.0"]
+build-backend = "uv_build"
+
+[tool.uv]
+package = true
+
+[tool.uv.sources]
+ocelescope = { workspace = true }
