+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/__init__.py
@@ -1,0 +1,3 @@
+from ocelescope_backend.app.main import app
+
+__all__ = ["app"]
