+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/internal/model/plugin.py
