+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/internal/registrar.py
