+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/internal/registry/resource.py
