+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/main.py
@@ -1,0 +1,3 @@
+from ocelescope_backend.app.factory import create_app
+
+app = create_app()
