+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/routes/resources.py
