+++ src/backend/ocelescope-backend/src/ocelescope_backend/app/sse_manager.py
