--- src/backend/pyproject.toml
@@ -1,46 +1,0 @@
-[project]
-name = "ocelescope-backend"
-version = "0.1"
-description = ""
-authors = [
-  { name = "Görkem-Emre Öztürk", email = "goerkem.emre.oeztuerk@gmail.com" },
-  { name = "Raimund Hensen", email = "raimund.hensen@rwth-aachen.de" },
-]
-requires-python = ">=3.13"
-dependencies = [
-  "pm4py~=2.7.22",
-  "pint>=0.23,<0.24",
-  "pandas>=2.3.1",
-  "numpy>=2.3.2",
-  "cachetools>=6.1.0",
-  "pydantic>=2.11.7",
-  "pydantic-settings>=2.2.1,<3",
-  "fastapi[standard]>=0.136.0",
-  "uvicorn>=0.28.0,<0.29",
-  "ocelescope",
-  "orjson>=3.11.3",
-]
-
-[project.urls]
-Repository = "https://github.com/rwth-pads/ocelescope"
-
-[dependency-groups]
-dev = [
-  "jupyterlab>=4.0.7,<5",
-  "iso4>=0.0.2,<0.0.3",
-  "seaborn>=0.13.0,<0.14",
-  "matplotlib>=3.8.0,<4",
-  "pre-commit>=4.2.0,<5",
-  "typer>=0.24.1",
-]
-
-[tool.uv]
-package = false
-
-[tool.uv.sources]
-ocelescope = { workspace = true }
-
-
-[tool.ruff]
-fix = true
-lint.ignore = ["E731"]
